+++ test/unit/fuzz-tools.test.js
@@ -261,6 +261,7 @@ import * as adiposityV270 from '../../lib/adiposity-v270.js';
 import * as lipidsV271 from '../../lib/lipids-v271.js';
 import * as anthroV272 from '../../lib/anthro-v272.js';
 import * as metabolicV273 from '../../lib/metabolic-v273.js';
+import * as proteinsV274 from '../../lib/proteins-v274.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -466,6 +467,7 @@ const MODULES = {
   'lipids-v271.js': lipidsV271,
   'anthro-v272.js': anthroV272,
   'metabolic-v273.js': metabolicV273,
+  'proteins-v274.js': proteinsV274,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];